console: updated persistent console to maintain state
Updated how the persistent console is rendered to maintain the request
context provider between opening/closing or switching to notebooks. This
will ensure that the output of the last result is not lost. We still
render null until the first time you open the console, but afterward we
continue to render the providers so that state is not lost.