[Console]  Maintain state of persistent console (#183864)
## Summary

Updating how the console wrapper is rendered to keep the request context
provider rendered once the console has been opened. This will allow
maintaining the output between sessions or switching between notebooks
and the console. But still allow unmounting the editor so that resources
are cleaned up when it's not in use.



https://github.com/elastic/kibana/assets/1972968/ae6b8ead-36f2-45c4-8571-f897fce94791