[SecuritySolution] Fix for timeline reset after active tab change (#183994)
## Summary

Fixes https://github.com/elastic/kibana/issues/181666

The timeline would reset after changing to the ESQL tab due to a
location update from the embedded Discover component. This adds a
workaround on our end to prevent the reset. A more detailed
investigation can be found in the linked issue above.