[CI] Serverless Release assistant to consider recent QAF builds (#184306)
## Summary
Builds on: https://github.com/elastic/qaf-tests/pull/75

If we only rely on the 3-hour cycle of the QAF test suite as an
indication for a "releaseable" build, then we always work with this 3-4
hour lead time. This can make releasing on-demand very delayed.

@pheyos added `PRE_RELEASE_CHECK=true` to some builds, we can use this
as an indicator that a test run was meant to prove the releaseability of
a version.

This PR also moves the release assistant job together with other
ES_Serverless jobs.