Update constructor for new API
These arguments changed in
https://github.com/elastic/kibana/pull/180741.

It looks like the default value for `allowNoIndices` (which was the
`true` positional arg we were passing previously passing) now defaults
to `false`, but in that PR I saw other equivalent calls being changed to
the form we see here. It seems to work, but I need to validate that the
behavior here really isn't changed.