[Detection Engine][Cypress] - Reduce exceptions flake - alerts_table_flow (#184383)
## Summary

Reduce flake for tests that check exceptions auto populate with alert
data. I moved to primarily checking what is being sent to the API after
the expected user interaction as a more solid way of confirming that the
flow is working as expected. Before we were checking that every value we
expect is being shown in the UI, which leads to flake depending on if
there are unexpected re-renders. There certainly are improvements to
make code side vs. test side, but this is a way to get our tests back up
and running.