[Security Solution] Rename `rulesManagementClient` -> `detectionRulesClient` (#184598)
**Partially addresses: https://github.com/elastic/kibana/issues/184364**

## Summary
As a first step in refactoring our newly added `rulesManagementClient`,
we are renaming it from `rulesManagementClient` to
`detectionRulesClient`.

`rulesManagementClient` looks and sounds too similar to a long existing
`rulesClient`, so it can be confusing to new devs or anyone quickly
digging through the code.

This PR just updates the name. Other refactorings will be arriving in
the upcoming PRs.