Fix type error in helper fn
Now that all of our rule types have a definition for
`alert_suppression`, a type error was being thrown from this helper
function. I observed the following facts:

1. Removing a definition of `alert_suppression` from _any_ rule type
   fixed the error
2. Removing ThresholdRule (with its different definition of
   `alert_suppression`) also fixed the error

So it seemed as though the z.discriminatedUnion worked if there were
either 1 or 3 definitions for `alert_suppression` (none, threshold, and
regular), but not for 2 definitions (threshold and regular).

After conferring with @marshallmain, this seems to be due to the use of
`Partial` in the return type of this function. Using `Omit` instead (and
reusing the return type from the inner helper fn) makes typescript
happy.