Warn if some specified index patterns do not contain matching indexes
* Changes similar "warning" utilities to have a common(ish) interface
* Updates some GLOBAL RULE LOGIC to carry through earlier warning
  messages that would have previously been overwritten
* Adds a single test for the current case

I'm a little wary of adding the `||=` logic to the rule executor, as it
almost certainly changes behavior somewhere. We'll see if any existing tests agree.

Reading through the code I'm trying to understand how this is _supposed_
to work at a high level; what is the relationship between these
different warnings? Some seem to be collected in groups, while others
are overwritten by subsequent checks. The latter may be bugs, I'll need
to discuss with the rest of the team.