[Lens] Fix reference line width stale update (#184414)
## Summary

Fixes #182775

Prevents stale update on line width update.
I suspect the root problem there was due to this commit
0657a0865187e28a1943bbaa4faa1167c17f1aba applied for too many
re-renderings within the annotation panel. I've removed the fix there
and used a `useMemo` for the `currentConfig` prop in the parent
component to stabilize the object reference on change.

I've profiled the renderings with the React profiler without noticing
any particular issue.