[Obs AI Assistant] Use const for function names (#184830)
Some of the AI Assistant functions have very generic names. One example
is the `context` function. When stumbling upon code like

```
const contextRequest = functionClient.hasFunction('context')
```

... it is quite difficult to navigate to the context function
implementation without knowing it's exact location. Searching for
`context` is futile since it's a term used for many different things.

Using a constant to refer to a function makes it much easier to navigate
to where the function is registered. It also avoids typos but that's a
side-effect, not the main motivation for this.

---------

Co-authored-by: kibanamachine <42973632+kibanamachine@users.noreply.github.com>