Provide ML fields on Define step of rule creation
This was previously only available on the About step, via the
useRuleIndices hook in combination with the useFetchIndex hook.

Add new composite hook that encapsulates the same logic, and provides it
to the define step. Unlike on the About step, we are currently only
using this for ML fields as other situations derive their field list
from a passed prop (which might be a performance optimization, or a bug,
or both).