Creates ES|QL where filters for ordinal charts (#184420)
## Summary

Closes https://github.com/elastic/kibana/issues/183425


![meow](https://github.com/elastic/kibana/assets/17003240/88baec65-5c35-41f8-b01e-bbce88536c7d)

This PR enables the creation of where clause filters by clicking a chart
in Discover and wherever the ES|QL editor exists. This means that this
is not available in dashboards.

This is possible only for ordinal charts. For date fields is quite
difficult to know the interval so I don't allow it for now. We already
support brushing so time filtering is already available.

### Checklist
- [ ] [Unit or functional
tests](https://www.elastic.co/guide/en/kibana/master/development-tests.html)
were updated or added to match the most common scenarios
- [ ] This was checked for [cross-browser
compatibility](https://www.elastic.co/support/matrix#matrix_browsers)

---------

Co-authored-by: kibanamachine <42973632+kibanamachine@users.noreply.github.com>
Co-authored-by: Lukas Olson <olson.lukas@gmail.com>