[APM] add view in apm button to alert details page (#184687)
## Summary
Relates to https://github.com/elastic/kibana/issues/184204
Relates to https://github.com/elastic/kibana/pull/184236

After adding the `View in APM` buttons to APM alert details
visualizations, a bug was introduced where the share plugin was not
available in the codepath.

This was due to using the `useAPMPluginContext` hook. APM plugin context
does not exist within the observability plugin where this component is
rendered.

The use of `useAPMPluginContext` did work when using the visualization
embeddables, because they are wrapped in context, but when using the
visualizations on their own, such as in the APM latency alert details
page, the experience was broken.

To fix the experience, we added some defensive code and hid the button
when the `share` plugin was not available.

Now, we want to introduce a permenant fix to re-introduce the View in
APM button the APM alert details visualization. This fix uses
`useKibana` rather than `useAPMPluginContext` to gain access to the
`share` plugin.

SLO APM latency SLI alert details

![image](https://github.com/elastic/kibana/assets/11356435/86fa3c4d-c0aa-4cf6-a9f1-46b8174c36d1)

SLO APM error rate SLI alert details

![image](https://github.com/elastic/kibana/assets/11356435/a519606b-bb6f-49c5-9c91-733e5c466390)

APM latency alert details 

![image](https://github.com/elastic/kibana/assets/11356435/1c9f2037-869d-4a57-836b-63c430ca9ee5)


### Testing

Both the APM SLO burn rate rules and the APM latency rules should be
tested in this PR

1. Generate APM test data. For example, you can use `node
scripts/synthtrace simple_trace --live`
2. Create an APM latency rule with the threshold that will trigger an
alert
3. Navigate to the alert details page
4. Ensure the page loads and the `View in APM` button appears
5. Click on the View in APM button to ensure it navigates to APM
appropriately
6. Create an APM latency SLI within the SLO app. Choose a threshold that
will trigger a low SLI value
7. Wait for a burn rate alert to fire (the rule is created
automatically)
8. Navigate to the alert details page. Ensure the page loads and the
`View in APM` button appears
9. Click on the View in APM button to ensure it navigates to APM
appropriately
10. Repeat steps 6 to 9 with the APM error rate SLI within the SLO app.