WIP: adding cypress tests for ML warning
To do this we need to get the ML API to recognize our jobs as installed
and running. They are currently _not_ recognizing this (although there
are anomalies in the index).

Still troubleshooting to see what's missing, here. This logic was
cribbed from the analogous FTR tests, but those also aren't working so
*shrug*.