Install test ML jobs with a known/compatible group
Specifying the `groups` parameter when using the "setup module" API
causes the corresponding jobs to be installed _with only the specified
group_. This meant that in our FTR tests, we have been installing jobs
with the `auditbeat` group.

However, part of the contract between ML and Detection Engine is that we
use the `group` parameter to determine relevance: if it doesn't belong
to either the `security` or (legacy) `siem` group(s), it effectively
does not exist to the Detection Engine.

This fixes the (very confusing) issue of jobs being installed but not
recognized, by specifying a recognized group id (and using our shared
constant for it), both in the FTR and cypress utilities.