Replace display names with actual Job IDs
The way the cypress helper that consumes these is written, both of these
forms work, but we're never going to encounter a rule with the display
name params, and knowing the display name but not the ID is not useful
for investigative purposes.

I can see how this might have been done to prevent needing to change
these jobs as their IDs change, but I think it's more likely that those
will change than their IDs.