Extract ML rule validation logic into hook
There's not a lot here, but I feel bad for adding anything to
step_define_rule so this is an attempt to minimize that.

In the course of refactoring I also caught a bug (perhaps just a test
environment one) where the form fields are temporarily `undefined` when
the hooks are run. I updated the form type to reflect this; hopefully
that doesn't have broader impact (but if it does, those are probably
also uncaught bugs).