[ES|QL] validate inline casting (#185904)
## Summary
Adds basic validation for inline casting. To do this I had to create a
new type of AST node and update the validation logic.

I sort of snuck in a refactor of our column checking logic as well. It
just kind of happened 🤷‍♂️

Check out the validation tests for some good cases:
https://github.com/elastic/kibana/pull/185904/files#diff-89c4af0faedcf80d51cfb19ae397a5898c7293055b19284a94cb3f6a7cd4d071R1948

### Not covered yet

One thing this doesn't cover is that some casts only work for some types
of data. For example `123::cartesian_point` won't work... the value has
to be a string. This is because `123::cartesian_point` is actually
syntactic sugar for `to_cartesianpoint(123)`.

In https://github.com/elastic/elasticsearch/pull/109713 Nik is giving us
a table of the casts to their functions so that we can derive this data
and be a bit more accurate.


### Checklist

- [x] [Unit or functional
tests](https://www.elastic.co/guide/en/kibana/master/development-tests.html)
were updated or added to match the most common scenarios

---------

Co-authored-by: Stratoula Kalafateli <efstratia.kalafateli@elastic.co>
Co-authored-by: Kibana Machine <42973632+kibanamachine@users.noreply.github.com>