[Security Solution] `DetectionRulesClient`: return `RuleResponse` from `createCustomRule` and `createPrebuiltRule` (#185748)
**Partially addresses: https://github.com/elastic/kibana/issues/184364**

## Summary
This PR changes `createCustomRule` and `createPrebuiltRule` methods to
return `RuleResponse` instead of `RuleAlertType`.

This is a continuation of the effort to improve `DetectionRulesClient`.
As a part of it we want to make the client return `RuleResponse` from
all its public methods. This is good because it lets us hide rule's
internal structure and centralise conversions in one place – in the
client. So in this and upcoming PRs we are going to convert
`RuleAlertType` to `RuleResponse` within the client's methods and return
the `RuleResponse` object.