search_homepage: skip ftrs for MKI
At least locally I was not able to get the roles working to use the
supertest without auth to enable/disable the feature flag for the FTR
tests. Given that I'm going to go ahead and skip these in MKI for now
and we'll have to test them before enabling the feature flag.