Update aliases on our v3 job index
These are obviously used to read/write from the anomalies index, so it
would make sense that things are behaving oddly if they're not correct.

NB that this is only fixed for the one v3 job, not the other; this
archive actually contains anomalies and mappings for two different
indexes.

Since the errors (and assertions) have thus far only concerned the
former, I'm not yet worrying about the latter.