Fix mappings/aliases for v3 ML job archive
These changes are taken from my investigation in #182183. Note that this
only changes this configuration for one of the ML jobs, but there are
two contained in this archive.

Since the tests are currently only concerned with the first job, so am
I. If in the course of investigation I find that we're not using the
other job's anomalies, I'll be deleting them since it's just test
overhead (and potentially a source of more bugs).