Stop ALL potential jobs at the start of our ML tests
I believe this is the cause of some sporadic failures when running these
tests together, as they all now have the implicit "start job when rule
is created" functionality.

By stopping the datafeeds before each suite, none of the relevant jobs
should be in the full "started" state required by our validation, so we
effectively have a "clean" state.

In figuring the above out I was also made aware of the "force stop
datafeed and close job" endpoint, which would probably be a more robust
solution, but for now this works.