Fix the body documentation
The route code intended the response to be `persistNote: res`. `res` was always undefined and therefore never returned. I changed the docs and the route code to match the behaviour.
Ideally we would have removed the `data: {}` object here but that would create a new route version which seems overkill.