[Dataset Quality] Fix flaky tests for Dataset Quality Summary (#184640)
## Summary

As part of the PR, i have rewritten 4 of the test files for both
serverless and stateful for Dataset Quality Project

- `/dataset_quality/dataset_quality_summary.ts`
  - Closes - https://github.com/elastic/kibana/issues/178874
  - Closes - https://github.com/elastic/kibana/issues/178884
  - Closes - https://github.com/elastic/kibana/issues/186354
- `/dataset_quality/dataset_quality_table.ts`
- Closes - https://github.com/elastic/kibana/issues/183940 (Possibly,
not guaranteed)
  - Closes - https://github.com/elastic/kibana/issues/182353
- `/dataset_quality/dataset_quality_table_filters.ts`
  - Closes - https://github.com/elastic/kibana/issues/183861
  - Closes - https://github.com/elastic/kibana/issues/182320
  - Closes - https://github.com/elastic/kibana/issues/184852
- `/dataset_quality/dataset_quality_flyout.ts`
  - Closes - https://github.com/elastic/kibana/issues/184438
  - Closes - https://github.com/elastic/kibana/issues/183851
  - Closes - https://github.com/elastic/kibana/issues/183771
  - Closes - https://github.com/elastic/kibana/issues/183525
  - Closes - https://github.com/elastic/kibana/issues/183312
  - Closes - https://github.com/elastic/kibana/issues/183129
  - Closes - https://github.com/elastic/kibana/issues/182154
  
## Why are the tests re-written

- Most of the `it` were loading its own data, which add 2 problems, 1.
Makes our tests slower, 2. Data cleanup becomes challenging. Now the
tests simply load one master set of data and all the Functional tests
can be executed on that master data. This makes our tests leaner and
more functional.

- Every `it` resets the page state after the tests. Like when a `it`
blocks opens the Flyout, it should also close the flyout which was
missing. In order to refresh the page, the `navigate` API was used,
which is not good. Navigate API should only be used once to navigate to
the page in the starting and then refresh events should be used if a
refresh is required, or the action should be un-done in order to get the
same state as previous. For ex - Sorting make update the state of the
whole page. At the end of the sorting test, sorting should be reset.
With these changes `it` block now only focus on pure functional testing.
This means the `it` blocks can be moved around, skipped without
impacting other tests

- We had too much of generic tests, which could be combined into 1 `it`
block and be checked together. Idea to split 1 `it` block into another
is when we test for a completely different scenario. For eg - Writing a
single `it` for testing different columns of a table is much more
cleaner than multiple `it` for testing various columns of the same
table.

- Removed usage of `retry.try`. (Personal Opinion, please read it with a
pinch of salt) - The retry service seems like an escape hatch (read
workaround) when we don't have control over the rendering of UI
elements. Better alternative is to use `retry.tryforTime` as the last
resort.
Also the only time i found using the whole `retry` package was when we
use the `browser` package for getting URL value of refreshing page. I
cannot prove yet the problem with the `browser` package but somehow it
breaks the sync behaviour causing elements to be not available hence
requiring these retries.
I have removed `browser.refresh` completely from our code in favour of
better refresh handlers using DateTimePicker Refresh action


Linked Issued - https://github.com/elastic/kibana/issues/184145