Reset combobox state after every option "fill"
The issue that the previous method (treating the first item different
from the rest) seems to be a consequence of there being "state" in the
dropdown list after the first item is selected. By hitting ESC after
each item is typed/selected, we get rid of this state, and so we can
treat each item separately again.