[Rollups] Add attributes for tracking doc links clicks (#186612)
Closes https://github.com/elastic/kibana/issues/186610

## Summary

This PR adds `data-test-subj` attributes to the doc links in the Rollup
deprecation warning callout and the deprecation empty callout so that
clicks on these links can be tracked on Fullstory. The links on the
deprecation callout have test subjects with a prefix depending on
whether they are on the list view page or the create form page so that
we can differentiate the clicks from the different pages.