[Security Solution][Endpoint] Bug fixes to the Endpoint List (#186223)
## Summary

Fixes a display bug on the Endpoint List where the Policy name column
value was wrapping and causing the row to be miss-aligned. Changes
include:

- Moved components pieces that display the Policy Revision and the
"Out-of-date" message to the `<EndpointPolicyLink>` component
- this component now handles on displaying all of this information in
one place via input Props
- The component also dues Authz checks and ensures that if the user does
not have Authz to read the Endpoint Policy Management section, the
component will display the policy name as plain text (no link)
- It will truncate the Policy name if not enough width is available to
display its full value
- Replaced the Policy List column component for Policy Name with the use
of `<EndpointPolicyLink>`
- Replaced the Policy Details flyout component to also use
`<EndpointPolicyLink>` to display the policy name

> [!NOTE]
> Its still possible for the Policy Name column on the Endpoint list to
display across two lines - when the Policy that the Endpoint host last
reported is not longer available in Kibana. In this case/flow, the
second line will display a message indicating that. See screen captures
below.