Disable ML suppression fields until they're loaded and available
It was previously possible to get into a state where the fields were not
loading, but were also an empty array. Since the test simply waits for
the field to be enabled, this meant there was a chance it would attempt
to fill unavailable options in the combobox. This removes that
possibility by ensuring that fields are present before enabling the
field.