[APM] Disable popover for grouped edges in Service Map (#186597)
Closes https://github.com/elastic/sdh-apm/issues/1378
Closes https://github.com/elastic/kibana/issues/186101

## Summary

Looking at the code, it seems like use-case for showing popover on an
edge that leads into a grouped resources node was never supported, and
it's unclear what kind of content should be displayed in the popover in
that case.

So this change just prevents the crashing reported in SDH by disabling
popover on grouped edged. In case we want to support this use-case,
there should a product decision about the content to show in the
popover.

https://github.com/elastic/kibana/assets/793851/9a353843-fd70-4543-a0b3-2fb8293d497f

Co-authored-by: Kibana Machine <42973632+kibanamachine@users.noreply.github.com>
(cherry picked from commit 17b010c603d3e364a75cba1472a7179dea647aa3)